TST: Split test_excel.py (#24472)
